add global test task
