--- packages/denodb/CHANGELOG.md
@@ -1,56 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [2.4.1](https://github.com/grammyjs/storages/compare/v2.4.0...v2.4.1) (2023-12-03)
-
-**Note:** Version bump only for package denodb
-
-# [2.4.0](https://github.com/grammyjs/storages/compare/v2.3.2...v2.4.0) (2023-10-09)
-
-**Note:** Version bump only for package denodb
-
-## [2.3.2](https://github.com/grammyjs/storages/compare/v2.3.1...v2.3.2) (2023-09-05)
-
-**Note:** Version bump only for package denodb
-
-## [2.3.1](https://github.com/grammyjs/storages/compare/v2.3.0...v2.3.1) (2023-08-28)
-
-**Note:** Version bump only for package denodb
-
-# [2.3.0](https://github.com/grammyjs/storages/compare/v2.2.0...v2.3.0) (2023-05-31)
-
-**Note:** Version bump only for package denodb
-
-# [2.2.0](https://github.com/grammyjs/storages/compare/v2.1.4...v2.2.0) (2023-04-15)
-
-**Note:** Version bump only for package denodb
-
-## [2.1.4](https://github.com/grammyjs/storages/compare/v2.1.3...v2.1.4) (2023-03-02)
-
-### Bug Fixes
-
-- **denodb:** again change denodb dependency to another url because it's again broken ([14ad651](https://github.com/grammyjs/storages/commit/14ad65134702ac3b4948bbf9321f7f5faf39df93))
-
-## [2.1.3](https://github.com/grammyjs/storages/compare/v2.1.2...v2.1.3) (2023-03-02)
-
-**Note:** Version bump only for package denodb
-
-## [2.1.2](https://github.com/grammyjs/storages/compare/v2.1.1...v2.1.2) (2023-03-02)
-
-**Note:** Version bump only for package denodb
-
-## [2.1.1](https://github.com/grammyjs/storages/compare/v2.1.0...v2.1.1) (2023-02-25)
-
-**Note:** Version bump only for package denodb
-
-# [2.1.0](https://github.com/grammyjs/storages/compare/v2.0.2...v2.1.0) (2023-01-16)
-
-### Bug Fixes
-
-- **denodb:** pin `denodb` dependency to github version instead of released one ([7bc7095](https://github.com/grammyjs/storages/commit/7bc70954e8809c032b8b94300aa84fd7ae15d6ad))
-
-### Features
-
-- unpin grammy version from all packages ([f814e3e](https://github.com/grammyjs/storages/commit/f814e3e675c31e599cfaa1c93975e8dd55d23be6))
