--- packages/denodb/deno.json
@@ -1,4 +1,0 @@
-{
-  "name": "@grammyjs/storage-deno",
-  "exports": "./src/mod.ts"
-}
