--- packages/denodb/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "denodb",
-  "private": true,
-  "scripts": {
-    "test:deno": "deno test --unstable ./test/session.test.ts"
-  },
-  "version": "2.4.1"
-}
