--- packages/denodb/src/deps.ts
@@ -1,3 +1,0 @@
-export type { StorageAdapter } from 'https://lib.deno.dev/x/grammy@1.x/mod.ts';
-// export * as DenoDB from 'https://lib.deno.dev/x/denodb@1.x/mod.ts';
-export * as DenoDB from 'https://raw.githubusercontent.com/jerlam06/denodb/master/mod.ts';
