--- packages/denodb/src/mod.ts
@@ -1,1 +1,0 @@
-export { DenoDBAdapter } from './adapter.ts';
