+++ packages/mongodb/__tests__/deno/deno.ts
@@ -1,16 +1,29 @@
-import { session } from 'grammy';
+import { Collection, MongoClient } from 'mongodb';
 import { test } from 'jsr:@std/testing/bdd';
 import { expect } from 'jsr:@std/expect';
-import { RedisAdapter } from '../src/mod.ts';
-import { RedisMock } from './redisMock.ts';
+import { session } from 'grammy';
 import { createBot, createMessage } from '@grammyjs/storage-utils';
+import { MongoDBAdapter } from '../../src/mod.ts';
+
+const createMongoClient = async () => {
+	const client = new MongoClient(`mongodb://localhost:27017/testdb `);
+	await client.connect();
+
+	return client;
+};
+
+const clearCollection = (col: Collection<any>) => col.deleteMany({});
 
 test('Pizza counter tests', async () => {
+	const client = await createMongoClient();
+	const db = client.db('testdb');
+	const collection = db.collection<any>('sessions');
+
 	const bot = createBot();
 
 	bot.use(session({
 		initial: () => ({ pizzaCount: 0 }),
-		storage: new RedisAdapter({ instance: new RedisMock() }),
+		storage: new MongoDBAdapter({ collection }),
 	}));
 
 	bot.hears('first', (ctx) => {
@@ -24,14 +37,23 @@ test('Pizza counter tests', async () => {
 
 	await bot.handleUpdate(createMessage(bot, 'first').update);
 	await bot.handleUpdate(createMessage(bot, 'second').update);
+
+	await clearCollection(collection);
+	client.close();
 });
 
 test('Simple string tests', async () => {
+	const client = await createMongoClient();
+	const db = client.db('testdb');
+	const collection = db.collection<any>('sessions');
+
 	const bot = createBot(false);
 
 	bot.use(session({
-		initial: () => 'test',
-		storage: new RedisAdapter({ instance: new RedisMock() as any }),
+		initial() {
+			return 'test';
+		},
+		storage: new MongoDBAdapter({ collection }),
 	}));
 
 	bot.hears('first', async (ctx) => {
@@ -44,4 +66,7 @@ test('Simple string tests', async () => {
 
 	await bot.handleUpdate(createMessage(bot, 'first').update);
 	await bot.handleUpdate(createMessage(bot, 'second').update);
+
+	await clearCollection(collection);
+	client.close();
 });
