+++ packages/psql/deno.json
@@ -3,7 +3,7 @@
   "exports": "./src/mod.ts",
   "version": "2.4.2",
   "tasks": {
-    "test": "deno test ./__tests__/deno.ts --unstable --no-check --allow-net --allow-import --allow-env"
+    "test": "deno test ./__tests__/deno/deno.ts --unstable --no-check --allow-net --allow-import --allow-env"
   },
   "imports": {
     "grammy": "https://lib.deno.dev/x/grammy@1.x/mod.ts",
